refactor(cli): split driver diagnostic helpers
